requirements.txt
# Coordinator dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-telegram-bot>=21.0
apprise>=1.9.0
kubernetes>=31.0.0
asyncpg>=0.30.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0
